refactor(database): column header
